+++ vck-openid/src/commonTest/kotlin/at/asitplus/wallet/lib/rqes/QtspAuthorizationTest.kt
@@ -50,7 +50,7 @@ class QtspAuthorizationTest : FreeSpec({
             ),
             authorizationDetails = credentialAuthReq.authorizationDetails
         )
-        qtspAuthenticationService.token(credentialTokenReq).getOrThrow()
+        qtspAuthenticationService.token(credentialTokenReq, null, null).getOrThrow()
     }
 
 })